Add Binary Sensor Entity
