+++ custom_components/miele/const.py
@@ -12,6 +12,7 @@ OAUTH_TOKEN_URL = "https://api.mcs3.miele.com/thirdparty/token"
 
 ENTITIES = [
     Platform.SENSOR,
+    Platform.BINARY_SENSOR,
 ]
 
 CAPABILITIES = {
